feat: add type discriminator
BREAKING CHANGE: Date objects now have a type
discriminator, in the form of a "type" field, which
is always set to the string "Date". The isDate and
isValidDate functions will not consider a value to
be a Date unless the discriminator is present. Most
other functions continue to accept objects without
this field, but if it is present it must be set to
the string "Date". This is to guard against errors
caused by similar objects such as DateTimes or
Timestamps being accidentally passed to date
functions. The type discriminator also makes it easy
to discriminate Dates from other types in type
unions.